requirements.txt
music21
python-midi
mido
python-rtmidi